# docs/API_OPENAPI_SCHEMA.yaml
openapi: 3.0.3
info:
  title: EDUZEN API
  description: API REST pour la plateforme de gestion scolaire EDUZEN
  version: 1.0.0
  contact:
    name: Support EDUZEN
    email: support@eduzen.com
servers:
  - url: https://app.eduzen.com/api
    description: Production
  - url: http://localhost:3001/api
    description: Development
...
paths:
...
  /2fa/verify-activation:
    post:
      tags: [2FA]
      summary: Vérifie un code TOTP lors de l'activation
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  pattern: '^\d{6}$'
                  example: "123456"
      responses:
        '200':
          description: 2FA activée avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /documents/generate:
    post:
      tags: [Documents]
      summary: Génère un document à partir d'un template
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateDocumentRequest'
      responses:
        '200':
          description: Document généré avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /compliance/alerts/check:
    post:
      tags: [Compliance]
      summary: Vérifie les alertes de conformité
      description: |
        Exécute toutes les vérifications d'alertes de conformité pour l'organisation de l'utilisateur.
        Les vérifications incluent GDPR, Qualiopi, et autres frameworks de conformité.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Résultats des vérifications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceCheckResponse'
              examples:
                success:
                  summary: Vérifications réussies
                  value:
                    success: true
                    timestamp: "2024-12-03T10:00:00Z"
                    results:
                      gdpr:
                        passed: true
                        alerts: []
                      qualiopi:
                        passed: true
                        alerts: []
                withAlerts:
                  summary: Vérifications avec alertes
                  value:
                    success: true
                    timestamp: "2024-12-03T10:00:00Z"
                    results:
                      gdpr:
                        passed: false
                        alerts:
                          - id: "alert-123"
                            type: "gdpr"
                            severity: "high"
                            message: "Données personnelles non chiffrées"
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    GenerateDocumentRequest:
      type: object
      required: [template_id, format]
      properties:
        template_id:
          type: string
        format:
          type: string
          enum: [pdf, docx, html]
        variables:
          type: object
        send_email:
          type: boolean
        email_to:
          type: string
          format: email
    DocumentResponse:
      type: object
      properties:
        success:
          type: boolean
        document_id:
          type: string
        file_url:
          type: string
          format: uri
        format:
          type: string
    ComplianceCheckResponse:
      type: object
      properties:
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        results:
          type: object
          additionalProperties:
            type: object
            properties:
              passed:
                type: boolean
              alerts:
                type: array
                items:
                  type: object
...
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Non authentifié
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    TooManyRequests:
      description: Trop de requêtes
      headers:
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: string
            format: date-time
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...